fix int cast

+++ RestaurantManagement/ViewModel/LichSuBanViewModel.cs
@@ -275,10 +275,10 @@ namespace QuanLyNhaHang.ViewModel
             ListProduct.Clear();
             while (reader.Read())
             {
-                int madon = reader.GetInt32(0);
+                int madon = reader.GetInt16(0);
                 string mamon = reader.GetString(1);
                 string ten = reader.GetString(2);
-                int soluong = reader.GetInt32(3);
+                int soluong = reader.GetInt16(3);
                 string gia = reader.GetSqlMoney(4).ToString();
                 string thoigian = reader.GetDateTime(5).ToShortDateString();
 
@@ -379,7 +379,7 @@ namespace QuanLyNhaHang.ViewModel
                             int madon = reader.GetInt16(0);
                             string mamon = reader.GetString(1);
                             string ten = reader.GetString(2);
-                            int soluong = reader.GetInt32(3);
+                            int soluong = reader.GetInt16(3);
                             string gia = reader.GetSqlMoney(4).ToString();
                             string thoigian = reader.GetDateTime(5).ToShortDateString();
